Remove DIP in favor of devcontainers (#2)
I figured that DIP is kind of redundant now that we've got advanced editor support.

Devcontainers are a good tool for running with VSCode, Cursor, GitHub Codespaces, and perhaps some other tools.

They also come with a simple CLI. Perhaps it doesn't provide the best UX, but it's still a lot more standard than what we had going with dip

While dip gave us a pretty good run, it's time to move on.